requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.1
openpyxl>=3.1.2
python-dotenv>=1.0.0
loguru>=0.7.2
numpy>=1.24.0